添加textMeshPro插件 添加方正字体 增加公告栏模块 增加消息和邮件模块 增加恶人榜模块

+++ Assets/Scripts/Data/MapInfoData.cs
@@ -6,6 +6,6 @@ public class MapInfoData {
     public string[] islandNames;
     public MinesData[] mines;
     public long moneyBox;
-    public int producePerSecond;
+    public double producePerSecond;
     public int limit;
 }
